refactor: hoist chat components
