+++ src/shared/navigationState.ts
